+++ changelog.md
@@ -8,6 +8,7 @@ Changelog
 * Added support for registry:delete command, to delete events from the
   registry.
 * Added registry:list-entities command.
+* Added registry:list-events command.
 
 
 0.1.0 (2023-03-29)